requirements.txt
cloudpickle==0.8.1
colour==0.1.5
cycler==0.10.0
deap==1.2.2
future==0.17.1
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
numpy-indexed==0.3.5
pandas==0.24.2
pygmo==2.10
pymongo==3.7.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-igraph==0.7.1.post6
pytimeparse==1.1.8
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
